updated install.sh

+++ install.sh
@@ -2,6 +2,32 @@
 
 [ "$EUID" -ne 0 ] && { echo 'This script must be run as root' >&2 ; exit 1 ;}
 
+confirm() {
+  local default="${2:-n}"
+  local prompt="${1:-Are you sure?}"
+
+  default="${default,,}"
+
+  if [[ "$default" == "y" ]]; then
+    prompt+=" [Y/n] "
+  else
+    prompt+=" [y/N] "
+  fi
+
+  while true; do
+    read -r -p "$prompt" ans
+    ans=${ans,,}
+
+    [[ -z "$ans" ]] && ans="$default"
+
+    case "$ans" in
+      y|yes) return 0 ;;
+      n|no) return 1 ;;
+      *) echo "Please answer yes or no" >&2 ;;
+    esac
+  done
+}
+
 pushd .
 mkdir -pv /usr/{ports,share/2} /etc/2
 
@@ -17,13 +43,19 @@ if [ ! -e /usr/ports/main ]; then
 fi
 
 ln -sfv scripts bin
-ln -siv "$PWD"/config.toml /etc/2/
-ln -siv "$PWD"/exclusions.txt /etc/2/
-ln -siv "$PWD"/repo_priority.txt /etc/2/
 
-rustup toolchain install nightly || true
-if ! cargo +nightly build --release; then
-    : # wget 'RELEASE URL'
+confirm 'Replace config?' && ln -sfv "$PWD"/config.toml /etc/2/
+confirm 'Replace exclusions?' && ln -sfv "$PWD"/exclusions.txt /etc/2/
+confirm 'Replace repo priority?' && ln -sfv "$PWD"/repo_priority.txt /etc/2/
+
+if confirm 'Compile from source (y) or use precompiled binary (n)?'; then
+    if ! command -v rustup > /dev/null 2>&1; then
+        echo "You don't have rustup; using precompiled binary instead" >&2
+        : # wget 'RELEASE URL'
+    fi
+
+    rustup toolchain install nightly || true
+    cargo +nightly build --release
 fi
 
 cat << EOF > /usr/bin/2
